uploading and downloading working
You can now upload files on the resources page. The files are stored in
the 'upload' folder in the PHP version folder, and when the file is
uploaded an entry is created in a mysql table called "file", whose
columns are filename, added, and size, all of which are type VARCHAR.
You guys might need to change the mysql connection initiation to get it
to work with your databases.