Lots of changes, PLEASE READ
Don't attempt to use the site before you update your database with the
teamsource.sql I have sent out.
As would be expected with a website requiring a login, the site will now
always redirect you back to the splash page if you are not logged in (or
in other words if you have not received a cookie from logging in).
To log in and view the site, you must create a user with the sign up
tool. Then, you will get a cookie lasting one hour, which can be
regained by logging in at the top with the email and password you
registered with. NOTE: passwords are stored in plaintext, so please do
not enter actual passwords. I'll  run you guys through it if we need to.